Refacto situation
